Settings Dispatcher: move it to its own module
This avoids the circular import that was originally a problem, and
removes the "weird" SettingsDispatcher assignment to itself, while
still adding documentation to keep this and the other dispatchers
"linked".

Reference: https://github.com/avocado-framework/avocado/issues/3065
